refactor: modular arch
